Copy the `xpring-eng` Code of Conduct from Xpring-JS (#1)
This ensures that any public repos under the `xpring-eng` organization
will have a CODE_OF_CONDUCT.md file. Additionally, as long as no repos
implement their own CODE_OF_CONDUCT, it means we have a single source of
truth for the content of the CODE_OF_CONDUCT for all our public repos,
and a single place we will need to make updates.